Add initial: Order model, repository, service, controller, event, and listener.
Created the Order model with the fields: id, customerId, total, and items.
Created the repository for Order.
Created the service for Order.
Created the controller for Order.
Created the OrderItemEvent to map the order items.
Created the OrderCreatedListener to consume an order Event.
Created the OrderCreatedEvent to be consumed by the listener.